requirements.txt
fastapi
redis
types-redis
uvicorn
tensorflow
sqlmodel
pytest
Pillow
numpy
fastapi_mail
jose
torch
torchvision
python-dotenv
authlib
httpx
itsdangerous
pyjwt
asyncio
asyncpg
psycopg2-binary
sqlalchemy